C++: string shuffler: changed type of second CLI argument to unsigned long to avoid overflows.

+++ languages/cpp/apps/string_shuffler/src/lib/icli_argument_validator.hpp
@@ -19,7 +19,7 @@ namespace shuffler::lib
    struct CLIArgs
    {
       std::string string_to_shuffle;
-      unsigned int number_of_shuffles;
+      unsigned long number_of_shuffles;
    };
 
    class ICLIArgumentValidator {
